--- a/crates/swc/src/plugin.rs
+++ b/crates/swc/src/plugin.rs
@@ -68,7 +68,7 @@ struct RustPlugins {
 }
 
 impl RustPlugins {
-    #[tracing::instrument(level = "trace", skip_all, name = "apply_plugins")]
+    #[tracing::instrument(level = "info", skip_all, name = "apply_plugins")]
     #[cfg(feature = "plugin")]
     fn apply(&mut self, n: Program) -> Result<Program, anyhow::Error> {
         use std::{path::PathBuf, sync::Arc};
@@ -88,7 +88,7 @@ impl RustPlugins {
         if let Some(plugins) = &self.plugins {
             for p in plugins {
                 let span = tracing::span!(
-                    tracing::Level::TRACE,
+                    tracing::Level::INFO,
                     "serialize_context",
                     plugin_module = p.0.as_str()
                 );
@@ -114,7 +114,7 @@ impl RustPlugins {
                 drop(context_span_guard);
 
                 let span = tracing::span!(
-                    tracing::Level::TRACE,
+                    tracing::Level::INFO,
                     "execute_plugin_runner",
                     plugin_module = p.0.as_str()
                 );