--- a/crates/swc_ecma_transforms_compat/src/bugfixes/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/bugfixes/mod.rs
@@ -12,7 +12,7 @@ mod edge_default_param;
 mod safari_id_destructuring_collision_in_function_expression;
 mod template_literal_caching;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn bugfixes() -> impl Fold {
     chain!(
         async_arrows_in_class(),