--- a/crates/swc_ecma_transforms_compat/src/es2015/block_scoped_fn.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2015/block_scoped_fn.rs
@@ -4,7 +4,7 @@ use swc_ecma_utils::UsageFinder;
 use swc_ecma_visit::{as_folder, noop_visit_mut_type, Fold, VisitMut, VisitMutWith};
 use swc_trace_macro::swc_trace;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn block_scoped_functions() -> impl Fold + VisitMut {
     as_folder(BlockScopedFns)
 }