--- a/crates/swc_ecma_transforms_compat/src/es2015/instanceof.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2015/instanceof.rs
@@ -31,7 +31,7 @@ use swc_trace_macro::swc_trace;
 ///
 /// _instanceof(foo, Bar);
 /// ```
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn instance_of() -> impl Fold + VisitMut {
     as_folder(InstanceOf)
 }