--- a/crates/swc_ecma_transforms_compat/src/es2015/object_super.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2015/object_super.rs
@@ -13,7 +13,7 @@ struct ObjectSuper {
     extra_vars: Vec<Ident>,
 }
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn object_super() -> impl Fold + VisitMut {
     as_folder(ObjectSuper {
         extra_vars: Vec::new(),