--- a/crates/swc_ecma_transforms_compat/src/es2015/regenerator/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2015/regenerator/mod.rs
@@ -27,7 +27,7 @@ pub struct Config {
     pub import_path: Option<JsWord>,
 }
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn regenerator(config: Config, top_level_mark: Mark) -> impl Fold + VisitMut {
     as_folder(Regenerator {
         config,