--- a/crates/swc_ecma_transforms_compat/src/es2015/sticky_regex.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2015/sticky_regex.rs
@@ -21,7 +21,7 @@ use swc_trace_macro::swc_trace;
 /// ```js
 /// new RegExp("o+", "y")
 /// ```
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn sticky_regex() -> impl 'static + Fold + VisitMut {
     as_folder(StickyRegex)
 }