--- a/crates/swc_ecma_transforms_compat/src/es2015/template_literal.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2015/template_literal.rs
@@ -12,7 +12,7 @@ use swc_ecma_utils::{
 use swc_ecma_visit::{as_folder, noop_visit_mut_type, Fold, VisitMut, VisitMutWith};
 use swc_trace_macro::swc_trace;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn template_literal(c: Config) -> impl Fold + VisitMut {
     as_folder(TemplateLiteral {
         c,