--- a/crates/swc_ecma_transforms_compat/src/es2017/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2017/mod.rs
@@ -5,7 +5,7 @@ pub use self::async_to_generator::async_to_generator;
 
 pub mod async_to_generator;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn es2017(c: Config) -> impl Fold {
     async_to_generator(c.async_to_generator)
 }