--- a/crates/swc_ecma_transforms_compat/src/es2019/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2019/mod.rs
@@ -4,7 +4,7 @@ pub use self::optional_catch_binding::optional_catch_binding;
 
 mod optional_catch_binding;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn es2019() -> impl Fold {
     optional_catch_binding()
 }