--- a/crates/swc_ecma_transforms_compat/src/es2019/optional_catch_binding.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2019/optional_catch_binding.rs
@@ -5,7 +5,7 @@ use swc_trace_macro::swc_trace;
 
 struct OptionalCatchBinding;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn optional_catch_binding() -> impl Fold + VisitMut {
     as_folder(OptionalCatchBinding)
 }