--- a/crates/swc_ecma_transforms_compat/src/es2021/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2021/mod.rs
@@ -4,7 +4,7 @@ pub use self::logical_assignments::logical_assignments;
 
 mod logical_assignments;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn es2021() -> impl Fold + VisitMut {
     logical_assignments()
 }