--- a/crates/swc_ecma_transforms_compat/src/es2022/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2022/mod.rs
@@ -10,10 +10,8 @@ pub mod class_properties;
 pub mod private_in_object;
 pub mod static_blocks;
 
-#[tracing::instrument(level = "trace", skip_all)]
-pub fn es2022<C: Comments>(cm: Option<C>, config: Config) -> impl Fold {
 #[tracing::instrument(level = "info", skip_all)]
-pub fn es2022(config: Config) -> impl Fold {
+pub fn es2022<C: Comments>(cm: Option<C>, config: Config) -> impl Fold {
     chain!(
         static_blocks(),
         class_properties(cm, config.class_properties),