--- a/crates/swc_ecma_transforms_compat/src/es2022/static_blocks.rs
+++ b/crates/swc_ecma_transforms_compat/src/es2022/static_blocks.rs
@@ -7,7 +7,7 @@ use swc_trace_macro::swc_trace;
 
 struct ClassStaticBlock;
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn static_blocks() -> impl Fold + VisitMut {
     as_folder(ClassStaticBlock)
 }