--- a/crates/swc_ecma_transforms_compat/src/es3/member_expr_lits.rs
+++ b/crates/swc_ecma_transforms_compat/src/es3/member_expr_lits.rs
@@ -20,7 +20,7 @@ use swc_trace_macro::swc_trace;
 /// obj["const"] = "isKeyword";
 /// obj["var"] = "isKeyword";
 /// ```
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn member_expression_literals() -> impl Fold {
     MemberExprLit
 }