--- a/crates/swc_ecma_transforms_compat/src/es3/mod.rs
+++ b/crates/swc_ecma_transforms_compat/src/es3/mod.rs
@@ -11,7 +11,7 @@ mod prop_lits;
 mod reserved_word;
 
 /// Make output es3-compatible.
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 pub fn es3(preserve_import: bool) -> impl Fold {
     chain!(
         property_literals(),