--- a/crates/swc_plugin_runner/src/cache.rs
+++ b/crates/swc_plugin_runner/src/cache.rs
@@ -48,7 +48,7 @@ pub struct PluginModuleCache {
     instantiation_lock: Mutex<()>,
 }
 
-#[tracing::instrument(level = "trace", skip_all)]
+#[tracing::instrument(level = "info", skip_all)]
 fn create_filesystem_cache(filesystem_cache_root: &Option<String>) -> Option<FileSystemCache> {
     let mut root_path = if let Some(root) = filesystem_cache_root {
         Some(PathBuf::from(root))
@@ -106,7 +106,7 @@ impl PluginModuleCache {
     /// [This code](https://github.com/swc-project/swc/blob/fc4c6708f24cda39640fbbfe56123f2f6eeb2474/crates/swc/src/plugin.rs#L19-L44)
     /// includes previous incorrect attempt to workaround file read issues.
     /// In actual transform, `plugins` is also being called per each transform.
-    #[tracing::instrument(level = "trace", skip_all)]
+    #[tracing::instrument(level = "info", skip_all)]
     pub fn load_module(&self, binary_path: &Path) -> Result<Module, Error> {
         let binary_path = binary_path.to_path_buf();
         let mut inner_cache = self.inner.get().expect("Cache should be available").lock();
@@ -127,7 +127,7 @@ impl PluginModuleCache {
 
         let load_cold_wasm_bytes = || {
             let span = tracing::span!(
-                tracing::Level::TRACE,
+                tracing::Level::INFO,
                 "load_cold_wasm_bytes",
                 plugin_module = binary_path.to_str()
             );