<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

- closes https://github.com/swc-project/swc/issues/3863.

I tried various approach to preserve compile time trace level but only enable Level::TRACE for `span` only. However, it simply not seems possible. The reason I choose trace level at the moment is I was under assumption it's the only level tracing spans will try to use and would possible to enable it selectively. Both seems incorrect - first, trace level can be used for logging any span-less event, or other places. Secondly, as said above we can't opt in to enable span's trace level.

PR changes span's level into `info`. This is not a regression or defeating back: https://tracing.rs/tracing_attributes/attr.instrument.html# If we didn't choose level intentionally it would been `info` all the time and we won't hit this problem from beginning.

> instruments a function to create and enter a tracing [span](https://docs.rs/tracing/latest/tracing/span/index.html) every time the function is called.
> Unless overriden, a span with info level will be generated.

WIth this PR, only tracing `spans` will be recorded as info. Any other traces, logs, or events can use TRACE | DEBUG level and `@swc/core` release binary will not emit those at all.


**BREAKING CHANGE:**

<!--
If this PR introduces a breaking change, it must contain a notice for it to be included in the CHANGELOG. Add description or remove entirely if not breaking.

You may need to update `MIGRATION.md` for the breaking changes.
-->

**Related issue (if exists):**