@kdy1 2022-03-07
I profiled it locally.

PR:
```
test full_es2015                  ... bench:  80,136,808 ns/iter (+/- 7,580,107)
test full_es2016                  ... bench:  85,148,904 ns/iter (+/- 30,743,348)
test full_es2017                  ... bench:  87,605,962 ns/iter (+/- 16,405,189)
test full_es2018                  ... bench:  84,212,345 ns/iter (+/- 13,840,869)
test full_es2019                  ... bench:  85,074,516 ns/iter (+/- 12,216,482)
test full_es2020                  ... bench:  80,233,583 ns/iter (+/- 14,899,070)
test full_es3                     ... bench: 111,405,208 ns/iter (+/- 10,115,031)
test full_es5                     ... bench: 103,204,875 ns/iter (+/- 9,714,975)
test parser                       ... bench:     246,809 ns/iter (+/- 8,527)
```

Main:
```
test full_es2015                  ... bench:  62,123,141 ns/iter (+/- 5,155,507)
test full_es2016                  ... bench:  61,233,145 ns/iter (+/- 6,222,669)
test full_es2017                  ... bench:  61,224,904 ns/iter (+/- 6,650,323)
test full_es2018                  ... bench:  60,357,304 ns/iter (+/- 4,089,180)
test full_es2019                  ... bench:  60,029,079 ns/iter (+/- 4,770,494)
test full_es2020                  ... bench:  57,809,924 ns/iter (+/- 4,155,606)
test full_es3                     ... bench:  81,534,441 ns/iter (+/- 4,828,373)
test full_es5                     ... bench:  76,303,150 ns/iter (+/- 4,709,092)
test parser                       ... bench:     246,544 ns/iter (+/- 3,130)
```

So it makes swc 15%~20% slower.