@kwonoj 2022-03-07
What do you think? Go vs no-go?

Personally, I think it's acceptable as likely we'll optimize / remove unnecessary spans by improving this. Currently we eagerly expands across every transform visitor, but based on real codebase observation we can cut some obvious paths. 

It'll still add overheads to runtime always, tradeoff to giving user level profiling features to end users.