@kdy1 2022-03-07
But I think it's too much to go with. Maybe we can use `swcx` binary for profiling instead