@kwonoj 2022-03-07
In current number I agree it's somewhat lot. My expectation is we can reduce those number by optimizing spans.

Not supporting trace in `@swc/core` means all of programattic access cannot profile, like `swc-loader` or jest or anything else.