@kwonoj 2022-03-08
I got curious about numbers, so ran it locally:

Main
```
test full_es2015                  ... bench: 222,920,918 ns/iter (+/- 8,316,859)
test full_es2016                  ... bench: 222,653,579 ns/iter (+/- 6,525,643)
test full_es2017                  ... bench: 225,245,446 ns/iter (+/- 7,668,584)
test full_es2018                  ... bench: 220,806,417 ns/iter (+/- 6,522,804)
test full_es2019                  ... bench: 223,732,250 ns/iter (+/- 11,723,803)
test full_es2020                  ... bench: 140,139,454 ns/iter (+/- 5,810,154)
test full_es3                     ... bench: 252,597,317 ns/iter (+/- 12,477,837)
test full_es5                     ... bench: 246,717,748 ns/iter (+/- 11,739,469)
test parser                       ... bench:     419,726 ns/iter (+/- 26,972)
```

PR
```
test full_es2015                  ... bench: 220,594,474 ns/iter (+/- 5,399,636)
test full_es2016                  ... bench: 220,703,386 ns/iter (+/- 11,409,377)
test full_es2017                  ... bench: 220,279,684 ns/iter (+/- 7,768,127)
test full_es2018                  ... bench: 219,599,446 ns/iter (+/- 13,968,280)
test full_es2019                  ... bench: 217,291,666 ns/iter (+/- 9,412,638)
test full_es2020                  ... bench: 134,070,813 ns/iter (+/- 4,670,114)
test full_es3                     ... bench: 250,844,650 ns/iter (+/- 7,959,047)
test full_es5                     ... bench: 244,216,406 ns/iter (+/- 7,729,769)
test parser                       ... bench:     415,734 ns/iter (+/- 20,809)
```

For both of cases, cpu utilization doesn't go over 20~25% in general. I'm curious how does benchmark intenrally works and if we could have some deterministic numbers. 

For optimization side, along with my above suggestion to optimize span itself, the part I'm hoping to get is  upstream support for sampling (https://github.com/tokio-rs/tracing/discussions/1474#discussioncomment-1022805) . I believe having both will allow reasonable overhead to runtime