@kdy1 2022-03-08
For cpu core, `cargo bench` intentionally uses only one cpu core, so it can't go high