@kdy1 2022-03-10
Seems like it's quite better.

First run:

```

test full_es2015                  ... bench:  61,109,499 ns/iter (+/- 2,305,026)
test full_es2016                  ... bench:  60,443,100 ns/iter (+/- 2,172,061)
test full_es2017                  ... bench:  60,217,412 ns/iter (+/- 5,933,569)
test full_es2018                  ... bench:  59,623,216 ns/iter (+/- 2,349,659)
test full_es2019                  ... bench:  59,241,583 ns/iter (+/- 2,840,240)
test full_es2020                  ... bench:  57,035,162 ns/iter (+/- 2,649,858)
test full_es3                     ... bench:  80,850,679 ns/iter (+/- 2,534,178)
test full_es5                     ... bench:  77,287,612 ns/iter (+/- 7,292,843)
test parser                       ... bench:     245,276 ns/iter (+/- 3,550)
```

Second run:

```
test full_es2015                  ... bench:  61,017,462 ns/iter (+/- 3,549,847)
test full_es2016                  ... bench:  60,511,858 ns/iter (+/- 2,967,864)
test full_es2017                  ... bench:  60,331,804 ns/iter (+/- 6,185,472)
test full_es2018                  ... bench:  59,419,608 ns/iter (+/- 4,134,257)
test full_es2019                  ... bench:  59,293,133 ns/iter (+/- 2,414,775)
test full_es2020                  ... bench:  56,766,325 ns/iter (+/- 3,164,065)
test full_es3                     ... bench:  80,697,641 ns/iter (+/- 2,591,066)
test full_es5                     ... bench:  75,963,858 ns/iter (+/- 4,211,157)
test parser                       ... bench:     246,035 ns/iter (+/- 12,311)
```


I'll rebase this PR to compare with the main branch.